# app/backend/src/main/java/br/com/pucc/projetointegradorvi/models/dto/FixedTransactionDto.java
package br.com.pucc.projetointegradorvi.models.dto;

public class FixedTransactionDto {

    private String institution;

    private String type;

    private String value;

    private String startDate;

    private String endDate;

    private String indexName;

    private String taxValue;

    public FixedTransactionDto() {}
    
	public String getInstitution() {
		return institution;
	}

	public void setInstitution(String institution) {
		this.institution = institution;
	}

	public String getType() {
		return type;
	}

	public void setType(String type) {
		this.type = type;
	}

	public String getValue() {
		return value;
	}

	public void setValue(String value) {
		this.value = value;
	}

	public String getStartDate() {
		return startDate;
	}

	public void setStartDate(String startDate) {
		this.startDate = startDate;
	}

	public String getEndDate() {
		return endDate;
	}

	public void setEndDate(String endDate) {
		this.endDate = endDate;
	}

	public String getIndexName() {
		return indexName;
	}

	public void setIndexName(String indexName) {
		this.indexName = indexName;
	}

	public String getTaxValue() {
		return taxValue;
	}

	public void setTaxValue(String taxValue) {
		this.taxValue = taxValue;
	}

}